Refactored classes to individual files
