+++ algorithm_toolkit/algorithm_exception.py
@@ -1,0 +1,2 @@
+class AlgorithmException(Exception):
+    """Raise ATK-specific errors"""
